requirements.txt
awscli==1.6.2
backports.ssl-match-hostname==3.4.0.2
bcdoc==0.12.2
boto==2.33.0
botocore==0.73.0
certifi==2015.9.6.2
colorama==0.2.5
decorator==3.4.0
docutils==0.12
GraphLab-Create==1.6.1
GraphLab-Create-License==1.6.1
ipykernel==4.0.3
ipython==4.0.0
ipython-genutils==0.1.0
Jinja2==2.8
jmespath==0.5.0
jsonschema==2.4.0
jupyter-client==4.0.0
jupyter-core==4.0.6
librato-metrics==0.4.9
MarkupSafe==0.23
mistune==0.7.1
mixpanel-py==3.1.1
multipledispatch==0.4.8
nbconvert==4.0.0
nbformat==4.0.0
notebook==4.0.4
path.py==0.0.0
pexpect==3.3
pickleshare==0.5
prettytable==0.7.2
ptyprocess==0.4
pyasn1==0.1.9
Pygments==2.0.2
python-dateutil==2.4.2
pyzmq==14.7.0
requests==2.3.0
rsa==3.1.2
simplegeneric==0.8.1
six==1.10.0
sseclient==0.0.8
terminado==0.5
tornado==4.1
traitlets==4.0.0
web.py==0.37
wheel==0.26.0
